+++ Makefile
@@ -16,8 +16,10 @@ NAME	= push_swap
 CC		= gcc
 CFLAGS	= -Wall -Wextra -Werror
 FT		= ./42_libft/libft.a
-# NUM		= 5
-SRCS	= main.c push_swap.c moves.c moves2.c utils.c three_args.c five_args.c radix.c helper.c validation.c
+SRCS	= main.c push_swap.c \
+		push.c  rev-rotate.c  rotate.c swap.c \
+		three_args.c five_args.c radix.c\
+		helper.c validation.c utils.c
 OBJSDIR	= obj
 OBJS	= $(addprefix ${OBJSDIR}/, ${SRCS:%.c=%.o})
 
@@ -43,11 +45,10 @@ ${FT}:
 	@${MAKE} -C 42_libft
 
 clean:
-	${MAKE} clean -C 42_libft
+	rm -f valgrind_check/*
 	rm -rf ${OBJSDIR}
 
 fclean: clean
-	rm -rf 42_libft
 	rm -f ${NAME}
 
 re: fclean all
