+++ src/radix.c
@@ -6,7 +6,7 @@
 /*   By: lucdos-s <lukas.facchi@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/07 20:12:04 by lucdos-s          #+#    #+#             */
-/*   Updated: 2023/03/02 19:59:34 by lucdos-s         ###   ########.fr       */
+/*   Updated: 2023/03/04 00:14:56 by lucdos-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
