+++ src/rev-rotate.c
@@ -1,56 +1,70 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   three_args.c                                       :+:      :+:    :+:   */
+/*   rev-rotate.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lucdos-s <lukas.facchi@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/01/23 21:07:02 by lucdos-s          #+#    #+#             */
+/*   Created: 2023/03/03 15:08:56 by lucdos-s          #+#    #+#             */
 /*   Updated: 2023/03/04 00:14:56 by lucdos-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	three_args_factor(t_node *node)
+void	move_rr(t_node *node)
 {
 	int	top;
-	int	mid;
-	int	bottom;
-	int	sum;
+	int	temp;
 
-	sum = 0;
-	top = node->num;
-	mid = node->next->num;
-	bottom = node->next->next->num;
-	if (top > mid)
-		sum++;
-	if (top > bottom)
-		sum++;
-	if (mid > bottom)
-		sum = sum + 3;
-	return (sum);
+	top = top_value(*node);
+	temp = node->num;
+	node->num = top;
+	if (node->next)
+		top = node->next->num;
+	node = node->next;
+	while (node)
+	{
+		node->num = temp;
+		temp = top;
+		if (node->next)
+			top = node->next->num;
+		node = node->next;
+	}
 }
 
-void	three_args(t_pswap push_swap)
+void	rrb(t_pswap *push_swap)
 {
-	int		factor;
+	move_rr(*(push_swap->stackb));
+	printf("rrb\n");
+}
 
-	factor = three_args_factor(*(push_swap.stacka));
-	if (factor == 1)
-		sa(push_swap);
-	if (factor == 2)
-		ra(&push_swap);
-	if (factor == 3)
+void	rra(t_pswap *push_swap)
+{
+	move_rr(*(push_swap->stacka));
+	printf("rra\n");
+}
+
+void	multi_ra(t_pswap push_swap, int n)
+{
+	int	i;
+
+	i = 0;
+	while (i < n)
 	{
-		sa(push_swap);
 		ra(&push_swap);
+		i++;
 	}
-	if (factor == 4)
-		rra(&push_swap);
-	if (factor == 5)
+}
+
+void	multi_rrb(t_pswap push_swap, int n)
+{
+	int	i;
+
+	i = 0;
+	while (i < n)
 	{
-		sa(push_swap);
-		rra(&push_swap);
+		rrb(&push_swap);
+		i++;
 	}
 }
