+++ src/rotate.c
@@ -1,48 +1,62 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   moves2.c                                           :+:      :+:    :+:   */
+/*   rotate.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lucdos-s <lukas.facchi@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/01/17 16:28:34 by lucdos-s          #+#    #+#             */
-/*   Updated: 2023/03/02 14:56:30 by lucdos-s         ###   ########.fr       */
+/*   Created: 2023/03/03 15:07:43 by lucdos-s          #+#    #+#             */
+/*   Updated: 2023/03/04 00:14:56 by lucdos-s         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-void ra(t_pswap *push_swap)
+void	move_r(t_node *node)
+{
+	int	num;
+
+	num = node->num;
+	while (node->next)
+	{
+		node->num = node->next->num;
+		node = node->next;
+	}
+	node->num = num;
+}
+
+void	ra(t_pswap *push_swap)
 {
 	move_r(*(push_swap->stacka));
 	printf("ra\n");
 }
 
-void rb(t_pswap *push_swap)
+void	rb(t_pswap *push_swap)
 {
 	move_r(*(push_swap->stackb));
 	printf("rb\n");
 }
 
-void rrb(t_pswap *push_swap)
+void	multi_rb(t_pswap push_swap, int n)
 {
-	move_rr(*(push_swap->stackb));
-	printf("rrb\n");
+	int	i;
+
+	i = 0;
+	while (i < n)
+	{
+		rb(&push_swap);
+		i++;
+	}
 }
 
-void rra(t_pswap *push_swap)
+void	multi_rra(t_pswap push_swap, int n)
 {
-	move_rr(*(push_swap->stacka));
-	printf("rra\n");
-}
+	int	i;
 
-void sa(t_pswap push_swap)
-{
-	printf("sa\n");
-	move_s(*(push_swap.stacka));
-}
-void sb(t_pswap push_swap)
-{
-	printf("sb\n");
-	move_s(*(push_swap.stackb));
+	i = 0;
+	while (i < n)
+	{
+		rra(&push_swap);
+		i++;
+	}
 }
