+++ valgrind_check/100 números
@@ -1,0 +1,1095 @@
+==1698== Memcheck, a memory error detector
+==1698== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
+==1698== Using Valgrind-3.15.0 and LibVEX; rerun with -h for copyright info
+==1698== Command: ./push_swap 100 2 70 63 44 79 93 66 72 18 90 59 56 96 61 77 49 31 99 36 15 22 97 37 6 53 3 10 52 95 58 86 28 11 34 4 78 55 9 12 32 7 85 46 39 35 24 81 91 27 65 23 47 50 16 67 64 29 62 43 75 20 74 68 25 54 89 69 87 1 38 94 14 8 5 17 92 13 71 80 82 40 76 73 51 48 26 33 57 21 30 45 98 83 84 41 60 42 88 19
+==1698== 
+==1698== error calling PR_SET_PTRACER, vgdb might block
+pb
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+pb
+ra
+ra
+ra
+ra
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+ra
+ra
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+ra
+ra
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+ra
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+ra
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+pb
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+ra
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+pb
+ra
+ra
+ra
+ra
+ra
+ra
+ra
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+pb
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pb
+pb
+ra
+pb
+ra
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+pb
+pb
+ra
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+ra
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+pb
+pb
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+ra
+pb
+pb
+ra
+pb
+ra
+pb
+ra
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+ra
+ra
+pb
+ra
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pb
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+pa
+==1698== 
+==1698== HEAP SUMMARY:
+==1698==     in use at exit: 0 bytes in 0 blocks
+==1698==   total heap usage: 866 allocs, 866 frees, 14,732 bytes allocated
+==1698== 
+==1698== All heap blocks were freed -- no leaks are possible
+==1698== 
+==1698== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
