+++ valgrind_check/2 números
@@ -1,0 +1,14 @@
+==1672== Memcheck, a memory error detector
+==1672== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
+==1672== Using Valgrind-3.15.0 and LibVEX; rerun with -h for copyright info
+==1672== Command: ./push_swap 2 5
+==1672== 
+==1672== error calling PR_SET_PTRACER, vgdb might block
+==1672== 
+==1672== HEAP SUMMARY:
+==1672==     in use at exit: 0 bytes in 0 blocks
+==1672==   total heap usage: 5 allocs, 5 frees, 68 bytes allocated
+==1672== 
+==1672== All heap blocks were freed -- no leaks are possible
+==1672== 
+==1672== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
