+++ valgrind_check/3 números
@@ -1,0 +1,15 @@
+==1724== Memcheck, a memory error detector
+==1724== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
+==1724== Using Valgrind-3.15.0 and LibVEX; rerun with -h for copyright info
+==1724== Command: ./push_swap 2 3 1
+==1724== 
+==1724== error calling PR_SET_PTRACER, vgdb might block
+rra
+==1724== 
+==1724== HEAP SUMMARY:
+==1724==     in use at exit: 0 bytes in 0 blocks
+==1724==   total heap usage: 7 allocs, 7 frees, 600 bytes allocated
+==1724== 
+==1724== All heap blocks were freed -- no leaks are possible
+==1724== 
+==1724== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
