+++ valgrind_check/Numeros repetidos
@@ -1,0 +1,15 @@
+==1711== Memcheck, a memory error detector
+==1711== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
+==1711== Using Valgrind-3.15.0 and LibVEX; rerun with -h for copyright info
+==1711== Command: ./push_swap 1 2 3 4 5 6 7 8 9 10 1
+==1711== 
+==1711== error calling PR_SET_PTRACER, vgdb might block
+Error
+==1711== 
+==1711== HEAP SUMMARY:
+==1711==     in use at exit: 0 bytes in 0 blocks
+==1711==   total heap usage: 1 allocs, 1 frees, 48 bytes allocated
+==1711== 
+==1711== All heap blocks were freed -- no leaks are possible
+==1711== 
+==1711== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
